refactor: Move spec files near to the implementation
