+++ packages/cozy-client/src/withMutation.spec.jsx
@@ -1,9 +1,9 @@
 import React from 'react'
 import { shallow } from 'enzyme'
 
-import CozyClient from '../CozyClient'
-import CozyLink from '../CozyLink'
-import withMutation from '../withMutation'
+import CozyClient from './CozyClient'
+import CozyLink from './CozyLink'
+import withMutation from './withMutation'
 
 describe('withMutation', () => {
   const NEW_TODO = {
